// projects/app/src/app/interfaces/oanda.ts
type TradeState = 'OPEN' | 'CLOSED' | 'CLOSE_WHEN_TRADEABLE';
// ...
export interface Trade {
  /** The Trade's identifier, unique within the Trade's Account. */
  id: string;

  /** The Trade's Instrument. */
  instrument: string;

  /** The execution price of the Trade. */
  price: string;

  /** The date/time when the Trade was opened. */
  openTime: string;

  /** The current state of the Trade. */
  state: TradeState;

  /**
   * The initial size of the Trade. Negative values indicate a short Trade,
   * and positive values indicate a long Trade.
   */
  initialUnits: string;

  /**
   * The margin required at the time the Trade was created. Note, this is the
   * ‘pure' margin required, it is not the ‘effective' margin used that
   * factors in the trade risk if a GSLO is attached to the trade.
   */
  initialMarginRequired: string;

  /**
   * The number of units currently open for the Trade. This value is reduced
   * to 0.0 as the Trade is closed.
   */
  currentUnits: string;

  /** The total profit/loss realized on the closed portion of the Trade. */
  realizedPL: string;

  /** The unrealized profit/loss on the open portion of the Trade. */
  unrealizedPL: string;

  /** Margin currently used by the Trade. */
  marginUsed: string;

  /**
   * The average closing price of the Trade. Only present if the Trade has
   * been closed or reduced at least once.
   */
  averageClosePrice: string;

  /** The IDs of the Transactions that have closed portions of this Trade. */
  closingTransactionIDs: string[];

  /** The financing paid/collected for this Trade. */
  financing: string;

  /** The dividend adjustment paid for this Trade. */
  dividendAdjustment: string;

  /**
   * The date/time when the Trade was fully closed. Only provided for Trades
   * whose state is CLOSED.
   */
  closeTime: string;

  /** The client extensions of the Trade. */
  clientExtensions: ClientExtensions;

  /**
   * Full representation of the Trade's Take Profit Order, only provided if
   * such an Order exists.
   */
  takeProfitOrder: TakeProfitOrder;

  /**
   * Full representation of the Trade's Stop Loss Order, only provided if such
   * an Order exists.
   */
  stopLossOrder: StopLossOrder;

  /**
   * Full representation of the Trade's Trailing Stop Loss Order, only
   * provided if such an Order exists.
   */
  trailingStopLossOrder: TrailingStopLossOrder;
}
// ...
export interface Transaction {
  /**
   * The Transaction's Identifier.
   */
  id: string;

  /**
   * The Orders Instrument.
   */
  instrument: string;

  /**
   * The date/time when the Transaction was created.
   */
  time: string;

  /**
   * The ID of the user that initiated the creation of the Transaction.
   */
  userID: number;

  /**
   * The ID of the Account the Transaction was created for.
   */
  accountID: string;

  /**
   * The ID of the “batch” that the Transaction belongs to. Transactions in
   * the same batch are applied to the Account simultaneously.
   */
  batchID: string;

  /**
   * The Request ID of the request which generated the transaction.
   */
  requestID?: string;

  /**
   * The Type of the Transaction.
   */
  type: TransactionType | string;
}
// ...
export interface TransferFundsTransaction extends Transaction {
  /**
   * The amount to deposit/withdraw from the Account in the Account's home
   * currency. A positive value indicates a deposit, a negative value
   * indicates a withdrawal.
   */
  amount: string;

  /**
   * The reason that an Account is being funded.
   */
  fundingReason: FundingReason;

  /**
   * An optional comment that may be attached to a fund transfer for audit
   * purposes
   */
  comment: string;

  /**
   * The Account's balance after funds are transferred.
   */
  accountBalance: string;
}
// ...
export interface OrderFillTransaction extends Transaction {
  /**
   * The ID of the Order filled.
   */
  orderID: string;

  /**
   * The client Order ID of the Order filled (only provided if the client has
   * assigned one).
   */
  clientOrderID: string;

  /**
   * The name of the filled Order's instrument.
   */
  instrument: string;

  /**
   * The number of units filled by the OrderFill.
   */
  units: number;

  /**
   * The HomeConversionFactors in effect at the time of the OrderFill.
   */
  homeConversionFactors: HomeConversionFactors;

  /**
   * The price that all of the units of the OrderFill should have been filled
   * at, in the absence of guaranteed price execution. This factors in the
   * Account's current ClientPrice, used liquidity and the units of the
   * OrderFill only. If no Trades were closed with their price clamped for
   * guaranteed stop loss enforcement, then this value will match the price
   * fields of each Trade opened, closed, and reduced, and they will all be
   * the exact same.
   */
  fullVWAP: string;

  /**
   * The price in effect for the account at the time of the Order fill.
   */
  fullPrice: ClientPrice;

  /**
   * The reason that an Order was filled
   */
  reason: OrderFillReason;

  /**
   * The profit or loss incurred when the Order was filled.
   */
  pl: string;

  /**
   * The profit or loss incurred when the Order was filled, in the
   * Instrument's quote currency.
   */
  quotePL: number;

  /**
   * The financing paid or collected when the Order was filled.
   */
  financing: string;

  /**
   * The financing paid or collected when the Order was filled, in the
   * Instrument's base currency.
   */
  baseFinancing: number;

  /**
   * The financing paid or collected when the Order was filled, in the
   * Instrument's quote currency.
   */
  quoteFinancing: number;

  /**
   * The commission charged in the Account's home currency as a result of
   * filling the Order. The commission is always represented as a positive
   * quantity of the Account's home currency, however it reduces the balance
   * in the Account.
   */
  commission: string;

  /**
   * The total guaranteed execution fees charged for all Trades opened, closed
   * or reduced with guaranteed Stop Loss Orders.
   */
  guaranteedExecutionFee: string;

  /**
   * The total guaranteed execution fees charged for all Trades opened, closed
   * or reduced with guaranteed Stop Loss Orders, expressed in the
   * Instrument's quote currency.
   */
  quoteGuaranteedExecutionFee: number;

  /**
   * The Account's balance after the Order was filled.
   */
  accountBalance: string;

  /**
   * The Trade that was opened when the Order was filled (only provided if
   * filling the Order resulted in a new Trade).
   */
  tradeOpened: TradeOpen;

  /**
   * The Trades that were closed when the Order was filled (only provided if
   * filling the Order resulted in a closing open Trades).
   */
  tradesClosed: TradeReduce[];

  /**
   * The Trade that was reduced when the Order was filled (only provided if
   * filling the Order resulted in reducing an open Trade).
   */
  tradeReduced: TradeReduce;

  /**
   * The half spread cost for the OrderFill, which is the sum of the
   * halfSpreadCost values in the tradeOpened, tradesClosed and tradeReduced
   * fields. This can be a positive or negative value and is represented in
   * the home currency of the Account.
   */
  halfSpreadCost: number;
}
// ...
export enum TransactionFilter {
  ORDER = 'Order-related Transactions. These are the Transactions that create, cancel, fill or trigger Orders',
  FUNDING = 'Funding-related Transactions',
  ADMIN = 'Administrative Transactions',
  CREATE = 'Account Create Transaction',
  CLOSE = 'Account Close Transaction',
  REOPEN = 'Account Reopen Transaction',
  CLIENT_CONFIGURE = 'Client Configuration Transaction',
  CLIENT_CONFIGURE_REJECT = 'Client Configuration Reject Transaction',
  TRANSFER_FUNDS = 'Transfer Funds Transaction',
  TRANSFER_FUNDS_REJECT = 'Transfer Funds Reject Transaction',
  MARKET_ORDER = 'Market Order Transaction',
  MARKET_ORDER_REJECT = 'Market Order Reject Transaction',
  LIMIT_ORDER = 'Limit Order Transaction',
  LIMIT_ORDER_REJECT = 'Limit Order Reject Transaction',
  STOP_ORDER = 'Stop Order Transaction',
  STOP_ORDER_REJECT = 'Stop Order Reject Transaction',
  MARKET_IF_TOUCHED_ORDER = 'Market if Touched Order Transaction',
  MARKET_IF_TOUCHED_ORDER_REJECT = 'Market if Touched Order Reject Transaction',
  TAKE_PROFIT_ORDER = 'Take Profit Order Transaction',
  TAKE_PROFIT_ORDER_REJECT = 'Take Profit Order Reject Transaction',
  STOP_LOSS_ORDER = 'Stop Loss Order Transaction',
  STOP_LOSS_ORDER_REJECT = 'Stop Loss Order Reject Transaction',
  GUARANTEED_STOP_LOSS_ORDER = 'Guaranteed Stop Loss Order Transaction',
  GUARANTEED_STOP_LOSS_ORDER_REJECT = 'Guaranteed Stop Loss Order Reject Transaction',
  TRAILING_STOP_LOSS_ORDER = 'Trailing Stop Loss Order Transaction',
  TRAILING_STOP_LOSS_ORDER_REJECT = 'Trailing Stop Loss Order Reject Transaction',
  ONE_CANCELS_ALL_ORDER = 'One Cancels All Order Transaction',
  ONE_CANCELS_ALL_ORDER_REJECT = 'One Cancels All Order Reject Transaction',
  ONE_CANCELS_ALL_ORDER_TRIGGERED = 'One Cancels All Order Trigger Transaction',
  ORDER_FILL = 'Order Fill Transaction',
  ORDER_CANCEL = 'Order Cancel Transaction',
  ORDER_CANCEL_REJECT = 'Order Cancel Reject Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY = 'Order Client Extensions Modify Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Order Client Extensions Modify Reject Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY = 'Trade Client Extensions Modify Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Trade Client Extensions Modify Reject Transaction',
  MARGIN_CALL_ENTER = 'Margin Call Enter Transaction',
  MARGIN_CALL_EXTEND = 'Margin Call Extend Transaction',
  MARGIN_CALL_EXIT = 'Margin Call Exit Transaction',
  DELAYED_TRADE_CLOSURE = 'Delayed Trade Closure Transaction',
  DAILY_FINANCING = 'Daily Financing Transaction',
  RESET_RESETTABLE_PL = 'Reset Resettable PL Transaction',
}
// ...
export interface TradeOpen {
  /**
   * The ID of the Trade that was opened
   */
  tradeID: string;

  /**
   * The number of units opened by the Trade
   */
  units: number;

  /**
   * The average price that the units were opened at.
   */
  price: string;

  /**
   * This is the fee charged for opening the trade if it has a guaranteed Stop
   * Loss Order attached to it.
   */
  guaranteedExecutionFee: string;

  /**
   * This is the fee charged for opening the trade if it has a guaranteed Stop
   * Loss Order attached to it, expressed in the Instrument's quote currency.
   */
  quoteGuaranteedExecutionFee: number;

  /**
   * The client extensions for the newly opened Trade
   */
  clientExtensions: ClientExtensions;

  /**
   * The half spread cost for the trade open. This can be a positive or
   * negative value and is represented in the home currency of the Account.
   */
  halfSpreadCost: string;

  /**
   * The margin required at the time the Trade was created. Note, this is the
   * ‘pure' margin required, it is not the ‘effective' margin used that
   * factors in the trade risk if a GSLO is attached to the trade.
   */
  initialMarginRequired: string;
}
// ...
export interface HomeConversionFactors {
  /**
   * The ConversionFactor in effect for the Account for converting any gains
   * realized in Instrument quote units into units of the Account's home
   * currency.
   */
  gainQuoteHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any losses
   * realized in Instrument quote units into units of the Account's home
   * currency.
   */
  lossQuoteHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any gains
   * realized in Instrument base units into units of the Account's home
   * currency.
   */
  gainBaseHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any losses
   * realized in Instrument base units into units of the Account's home
   * currency.
   */
  lossBaseHome: ConversionFactor;
}

/**
 * A ConversionFactor contains information used to convert an amount, from an
 * Instrument's base or quote currency, to the home currency of an Account.
 */
export interface ConversionFactor {
  factor: number;
}